requirements.txt
discord.py[voice]>=1.7.3
boto3>=1.16.24
aiohttp>=3.7.4.post0